Added inspection for class-level attributes containing upper-class letters.
"noqa: Z214" is needed on WrongNameVisitor since I could not find a way to implement this without adding ClassDef-specific handler to the class, which in turn raised method count over the limit.

I could have maintaned a "last visited node" context inside the visitor and trigger check in standard place, but it would not be explicit enough in my opinion.