Fix `TooManyMethodsViolation` not displaying line number. (#233)
When `TooManyMethodsViolation` was triggered, flake8 showed it happening at line number 0. This was due to this violation being SimpleViolation, instead of a more appropriate ASTViolation. It may have occured because other violations in the same packageare module-level ones and therefore it is sufficient for them to be `simple`, without pointing to particular node.

I am not sure how to proceed with tests here, seeing that it may involve breaking current philosophy of using provided `assert_errors` helper, since it checks only class of the violation without looking at the node contents.