Issue 933 reference isort settings (#936)
* Fixed link to `isort.toml` in config docs

The link was supposed to point to `isort.toml`, but did actually point
to `flake8.toml`. This is fixed now.

* Updated docstring to reference `isort` settings

`isort` settings need to include `multi_line_output = 3` option to not
collide with WSP318. The docstring has been updated with a reference.

* Updated CHANGELOG

* Fixed which docstring is update

I messed up before and changed the docstring of WSP317 instead of 318.
This is fixed now.