revert: Annotations semantic checks
Why? Because it is too complex and bring a very little value.

Let me explain. Annotations can be used in multiple ways:

1. As Python structures: `List[int]`
2. As strings: `'SelfType[T]'`
3. As a mix of these two: `List['SelfType[T]']`

It is also really hard to tell that annotations are really the same.
Because, here's how one can write `Literal` (as an example):

1. `Literal[1]`
2. `typing.Literal[1]`
3. `t.Literal[1]`
4. `te.Literal[1]`
5. `typing_extensions.Literal[1]`
6. `'Literal[1]'`
7. ...

It would be the same annotation in all these cases. But different `ast` nodes.
Comparing them is hard. And error-prone.

We also break one very important princible of our project: we assume types.
That's why it is reverted.