Remove ImplicitTernaryViolation - WPS332 (#1099)
We are removing this rule because we found out that it was not
applicable to all cases.

This is the first time we remove an existing rule, so a mechanism to
keep track of them and satisfy test_no_holes was needed. The implemented
mechanism involves just adding the deprecated code to the
DEPRECATED_CODES tuple, which must be defined on the file where the
violation class lived. This variable is not defined at all if there are
no deprecated codes on the file.