Issue#22 (#24)
* test: added a new test that checks that methods named get/set are allowed if not containing the name of a class attribute or instance attribute.
relates to #22

* fix: changed description

* test: added a test for checking that correct use of dataclass with property annotation is valid.
relates to #22 and #1054

* test: added test with class attribute and instance getters/setters that should be prohibited.
relates to #22 and #1054

* test:
added new test cases for test_nonmatching_attribute_getter_setter, such that different
types of assignments are tested.
relates to #22 and #1054

* refactor:
refactored test_class_mixed to use stacked test parameters to make it easier to test
many combinations. Relates to #22 and #1054

* refactor:
added accessor as a new test parameter to ease the creation of
different test combinations for test_nonmatching_attribute_getter_setter and
test_instance_and_class_getter_setter.
Relates to #22 and #1054

* refactor:
renamed parameters in test_instance_and_class_getter_setter and test_nonmatching_attribute_getter_setter
to ease readability.
relates to #22 and #1054

* test:
added tests that checks that invalid use of class attributes
and getters and setters are prohbited.
Relates to #22, #1054

* test:
added test to ensure that class attributes with
non-matching getters and setters is allowed.
relates to #22, #1054

* test:
added a test for dataclass that uses explicit getters and setters on
an instance attribute. This should be prohibited. Right now the test
is failing, which indicates that something in the implementation needs to be looked at.
Relates to #22, #1054

* Feat: add checks for AnnAssign

* Refactor: rename variable ´asd´ to ´attribute´

* Test: Adds 3 tests

* Refactor: Renames some tests

* Refactor: Cleans up the tests

* Fix: test in wrong location

* Feat: seperate class and instance methods

* Fix: some linter errors

* Test: Moves test and readds deleted test

* test:
added a test that checks that correct use of dataclass with
the propert annotation is allowed
relates to #22 and #1054

* Fix: linter violations

Co-authored-by: fwald <38955346+fwald@users.noreply.github.com>
Co-authored-by: ruwaid123 <43616385+ruwaid123@users.noreply.github.com>