Git ignore asdf .tool-versions file
Asdf is a version manager that can be used for python and other programming
languages and tools. It is similar to pyenv in that regard. It uses a
`.tool-versions` file to specify versions of tools used in a project. Since
pyenv's equivalent, `.python-version`, is ignored, `.tool-versions` should be
ignored as well.