Redundant Tenary Check Added
Added RedundantTernaryViolation to best_practices.py with code 475. The error template for this code is set to 'Found redundant ternary operator'.
Added RedundantTernaryVisistor to redundancy.py. This visistor checks for all the cases mentioned in the original issue request.
Added The RedundantTernaryVisistor to the list of visitors in tree.py.
Added a test case for the RedundantTernaryVisistor to new file test_redundant_ternary.py.
Added 3 test cases to noqa.py to test the new check under code 475. Reflected this in test_noqa.py.